+++ CHANGELOG.md
@@ -7,22 +7,28 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Admin Cost Center addresses UI
+
 ## [1.35.2] - 2024-10-15
+
 ### Added
 
 - Add new tab called 'settings' at organization details
-
 
 ## [1.35.1] - 2024-10-01
 
 ## [1.35.0] - 2024-09-10
 
 ### Added
+
 - Added `getAccount` query on OrganizationDetailsSellers component to get the account data
 
 ## [1.34.3] - 2024-09-05
 
 ### Fixed
+
 - Review and reorder translations for better maintainability
 
 ## [1.34.2] - 2024-09-03
@@ -40,6 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.34.0] - 2024-08-20
 
 ### Fixed
+
 - Implement pagination for sellers in Organization Details page
 
 ## [1.33.1] - 2024-08-06
@@ -69,6 +76,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Implement 'none' option at trade policy organization
 
 ## [1.31.9] - 2024-06-20
+
 ### Fixed
 
 - Adjusts list organization and research screening according to input at autocomplete
@@ -80,6 +88,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.31.6] - 2024-05-28
 
 ## [1.31.5] - 2024-05-28
+
 ### Fixed
 
 - Save new address within the Cost Center in the store